+++ ShipmentGeek/frmAbout.designer.cs
@@ -27,16 +27,19 @@ namespace ShipmentGeek
         /// </summary>
         private void InitializeComponent()
         {
+            this.components = new System.ComponentModel.Container();
             this.tableLayoutPanel = new System.Windows.Forms.TableLayoutPanel();
             this.labelConfiguration = new System.Windows.Forms.Label();
             this.labelDevelopedBy = new System.Windows.Forms.Label();
             this.logoPictureBox = new System.Windows.Forms.PictureBox();
             this.labelProductName = new System.Windows.Forms.Label();
+            this.labelVersion = new System.Windows.Forms.Label();
             this.labelCopyright = new System.Windows.Forms.Label();
             this.labelCompanyName = new System.Windows.Forms.Label();
             this.textBoxDescription = new System.Windows.Forms.TextBox();
             this.okButton = new System.Windows.Forms.Button();
-            this.labelLibrary = new System.Windows.Forms.Label();
+            this.label1 = new System.Windows.Forms.Label();
+            this.toolTip = new System.Windows.Forms.ToolTip(this.components);
             this.tableLayoutPanel.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.logoPictureBox)).BeginInit();
             this.SuspendLayout();
@@ -50,11 +53,12 @@ namespace ShipmentGeek
             this.tableLayoutPanel.Controls.Add(this.labelDevelopedBy, 0, 4);
             this.tableLayoutPanel.Controls.Add(this.logoPictureBox, 0, 0);
             this.tableLayoutPanel.Controls.Add(this.labelProductName, 1, 0);
-            this.tableLayoutPanel.Controls.Add(this.labelLibrary, 1, 1);
+            this.tableLayoutPanel.Controls.Add(this.labelVersion, 1, 1);
             this.tableLayoutPanel.Controls.Add(this.labelCopyright, 1, 2);
             this.tableLayoutPanel.Controls.Add(this.labelCompanyName, 1, 3);
             this.tableLayoutPanel.Controls.Add(this.textBoxDescription, 1, 5);
             this.tableLayoutPanel.Controls.Add(this.okButton, 1, 6);
+            this.tableLayoutPanel.Controls.Add(this.label1, 0, 5);
             this.tableLayoutPanel.Dock = System.Windows.Forms.DockStyle.Fill;
             this.tableLayoutPanel.Location = new System.Drawing.Point(9, 9);
             this.tableLayoutPanel.Name = "tableLayoutPanel";
@@ -67,14 +71,18 @@ namespace ShipmentGeek
             this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 40F));
             this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Percent, 10F));
             this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
-            this.tableLayoutPanel.Size = new System.Drawing.Size(417, 265);
+            this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this.tableLayoutPanel.RowStyles.Add(new System.Windows.Forms.RowStyle(System.Windows.Forms.SizeType.Absolute, 20F));
+            this.tableLayoutPanel.Size = new System.Drawing.Size(417, 276);
             this.tableLayoutPanel.TabIndex = 0;
             // 
             // labelConfiguration
             // 
             this.labelConfiguration.Anchor = System.Windows.Forms.AnchorStyles.Left;
             this.labelConfiguration.ImageAlign = System.Drawing.ContentAlignment.MiddleRight;
-            this.labelConfiguration.Location = new System.Drawing.Point(6, 242);
+            this.labelConfiguration.Location = new System.Drawing.Point(6, 252);
             this.labelConfiguration.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
             this.labelConfiguration.MaximumSize = new System.Drawing.Size(128, 17);
             this.labelConfiguration.Name = "labelConfiguration";
@@ -86,7 +94,7 @@ namespace ShipmentGeek
             // labelDevelopedBy
             // 
             this.labelDevelopedBy.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.labelDevelopedBy.Location = new System.Drawing.Point(143, 104);
+            this.labelDevelopedBy.Location = new System.Drawing.Point(143, 108);
             this.labelDevelopedBy.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
             this.labelDevelopedBy.MaximumSize = new System.Drawing.Size(0, 17);
             this.labelDevelopedBy.Name = "labelDevelopedBy";
@@ -97,14 +105,18 @@ namespace ShipmentGeek
             // 
             // logoPictureBox
             // 
+            this.logoPictureBox.Cursor = System.Windows.Forms.Cursors.Hand;
+            this.logoPictureBox.Image = global::ShipmentGeek.Properties.Resources.CigarGeek;
             this.logoPictureBox.InitialImage = null;
             this.logoPictureBox.Location = new System.Drawing.Point(3, 3);
             this.logoPictureBox.Name = "logoPictureBox";
-            this.tableLayoutPanel.SetRowSpan(this.logoPictureBox, 6);
-            this.logoPictureBox.Size = new System.Drawing.Size(100, 143);
+            this.tableLayoutPanel.SetRowSpan(this.logoPictureBox, 5);
+            this.logoPictureBox.Size = new System.Drawing.Size(128, 128);
             this.logoPictureBox.SizeMode = System.Windows.Forms.PictureBoxSizeMode.AutoSize;
             this.logoPictureBox.TabIndex = 12;
             this.logoPictureBox.TabStop = false;
+            this.toolTip.SetToolTip(this.logoPictureBox, "Open CigarGeeks website");
+            this.logoPictureBox.Click += new System.EventHandler(this.logoPictureBox_Click);
             // 
             // labelProductName
             // 
@@ -119,10 +131,22 @@ namespace ShipmentGeek
             this.labelProductName.Text = "Product Name";
             this.labelProductName.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
             // 
+            // labelVersion
+            // 
+            this.labelVersion.Dock = System.Windows.Forms.DockStyle.Fill;
+            this.labelVersion.Location = new System.Drawing.Point(143, 27);
+            this.labelVersion.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
+            this.labelVersion.MaximumSize = new System.Drawing.Size(0, 17);
+            this.labelVersion.Name = "labelVersion";
+            this.labelVersion.Size = new System.Drawing.Size(271, 17);
+            this.labelVersion.TabIndex = 0;
+            this.labelVersion.Text = "Version";
+            this.labelVersion.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
+            // 
             // labelCopyright
             // 
             this.labelCopyright.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.labelCopyright.Location = new System.Drawing.Point(143, 52);
+            this.labelCopyright.Location = new System.Drawing.Point(143, 54);
             this.labelCopyright.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
             this.labelCopyright.MaximumSize = new System.Drawing.Size(0, 17);
             this.labelCopyright.Name = "labelCopyright";
@@ -134,7 +158,7 @@ namespace ShipmentGeek
             // labelCompanyName
             // 
             this.labelCompanyName.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.labelCompanyName.Location = new System.Drawing.Point(143, 78);
+            this.labelCompanyName.Location = new System.Drawing.Point(143, 81);
             this.labelCompanyName.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
             this.labelCompanyName.MaximumSize = new System.Drawing.Size(0, 17);
             this.labelCompanyName.Name = "labelCompanyName";
@@ -146,13 +170,13 @@ namespace ShipmentGeek
             // textBoxDescription
             // 
             this.textBoxDescription.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.textBoxDescription.Location = new System.Drawing.Point(143, 133);
+            this.textBoxDescription.Location = new System.Drawing.Point(143, 138);
             this.textBoxDescription.Margin = new System.Windows.Forms.Padding(6, 3, 3, 3);
             this.textBoxDescription.Multiline = true;
             this.textBoxDescription.Name = "textBoxDescription";
             this.textBoxDescription.ReadOnly = true;
             this.textBoxDescription.ScrollBars = System.Windows.Forms.ScrollBars.Both;
-            this.textBoxDescription.Size = new System.Drawing.Size(271, 100);
+            this.textBoxDescription.Size = new System.Drawing.Size(271, 104);
             this.textBoxDescription.TabIndex = 23;
             this.textBoxDescription.TabStop = false;
             this.textBoxDescription.Text = "Description";
@@ -162,29 +186,28 @@ namespace ShipmentGeek
             this.okButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
             this.okButton.DialogResult = System.Windows.Forms.DialogResult.Cancel;
             this.okButton.ImageAlign = System.Drawing.ContentAlignment.MiddleLeft;
-            this.okButton.Location = new System.Drawing.Point(339, 239);
+            this.okButton.Location = new System.Drawing.Point(339, 250);
             this.okButton.Name = "okButton";
             this.okButton.Size = new System.Drawing.Size(75, 23);
             this.okButton.TabIndex = 24;
             this.okButton.Text = "&OK";
             // 
-            // labelLibrary
+            // label1
             // 
-            this.labelLibrary.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.labelLibrary.Location = new System.Drawing.Point(143, 26);
-            this.labelLibrary.Margin = new System.Windows.Forms.Padding(6, 0, 3, 0);
-            this.labelLibrary.MaximumSize = new System.Drawing.Size(0, 17);
-            this.labelLibrary.Name = "labelLibrary";
-            this.labelLibrary.Size = new System.Drawing.Size(271, 17);
-            this.labelLibrary.TabIndex = 0;
-            this.labelLibrary.Text = "Version";
-            this.labelLibrary.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
+            this.label1.Location = new System.Drawing.Point(3, 135);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(131, 110);
+            this.label1.TabIndex = 27;
+            this.label1.Text = "ShipmentGeek is not affiliated with, endorsed by, or in any way associated with C" +
+    "igarGeeks. We simply want to inform you of the greatest cigar community on the w" +
+    "eb.";
+            this.label1.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
             // 
             // frmAbout
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(435, 283);
+            this.ClientSize = new System.Drawing.Size(435, 294);
             this.Controls.Add(this.tableLayoutPanel);
             this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedDialog;
             this.MaximizeBox = false;
@@ -213,6 +236,8 @@ namespace ShipmentGeek
         private System.Windows.Forms.Label labelDevelopedBy;
         private System.Windows.Forms.TextBox textBoxDescription;
         private System.Windows.Forms.Button okButton;
-        private System.Windows.Forms.Label labelLibrary;
+        private System.Windows.Forms.Label labelVersion;
+        private System.Windows.Forms.ToolTip toolTip;
+        private System.Windows.Forms.Label label1;
     }
 }