+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Show dialog box when removing shipment and shipment item.
 * Added search for shipments.
 * Added menu strip and about form.
+* Shipment file is selectable.
 
 #### v.1.2.0
 * Improved shipment selection actions.